Modify: SingleCardArea와 MultiCardArea를 병합
